backend/requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
google-generativeai==0.3.1
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
textblob==0.17.1
flask==3.0.0